+++ src/disk_manager/disk_manager.cc
@@ -7,6 +7,7 @@
 #include <string.h>
 #include <errno.h>
 #include <glog/logging.h>
+#include "../network_manager/request_data.h"
 
 namespace Embarcadero{
 
@@ -56,32 +57,32 @@ void DiskManager::Disk_io_thread(){
 		if(!optReq.has_value()){
 			break;
 		}
-		const struct PublishRequest &req = optReq.value();
-		int off = offset_.fetch_add(req.req->payload_size(), std::memory_order_relaxed);
-		DLOG(INFO) << "Received payload is: " << req.req->payload().c_str();
-		pwrite(log_fd_, req.req->payload().c_str(), req.req->payload_size(), off);
+		struct PublishRequest &req = optReq.value();
+		struct RequestData *req_data = static_cast<RequestData*>(req.grpcTag);
 
-		// Post I/O work (as disk I/O depend on the same payload)
+		int off = offset_.fetch_add(req_data->request_.payload_size(), std::memory_order_relaxed);
+		DLOG(INFO) << "Received payload is: " << req_data->request_.payload().c_str();
+		pwrite(log_fd_, req_data->request_.payload().c_str(), req_data->request_.payload_size(), off);
+
+		// TODO(erika): below logic should really be shared function between CXL and Disk managers
 		int counter = req.counter->fetch_sub(1, std::memory_order_relaxed);
 
 		// If no more tasks are left to do
 		if (counter == 0) {
-			if (req.req->acknowledge()) {
+			if (req_data->request_.acknowledge()) {
 				// TODO: Set result - just assume success
-				network_manager_->SetError(req.grpcTag, ERR_NO_ERROR);
+				req_data->SetError(ERR_NO_ERROR);
 
 				// Send to network manager ack queue
 				auto maybeTag = std::make_optional(req.grpcTag);
-				// TODO(erika) remove printf
-				printf("DiskManager enquing to ack queue, tag=%p\n", req.grpcTag);
+				DLOG(INFO) << "Enquing to ack queue, tag=" << req.grpcTag;
 				EnqueueAck(ackQueue_, maybeTag);
 			} else {
-				// gRPC has already sent response, so here we can just free the CallData object.
-				DLOG(INFO) << "DiskManager calling proceed on call data";
-				network_manager_->Proceed(req.grpcTag);
+				// gRPC has already sent response, so just mark the object as ready for destruction
+				req_data->Proceed();
 			}
 		} else {
-			DLOG(INFO) << "DiskManager got counter: " << counter;
+			DLOG(INFO) << "counter: " << counter;
 		}
 	}
 }
