+++ src/disk_manager/disk_manager.h
@@ -6,19 +6,15 @@
 #include <optional>
 #include "folly/MPMCQueue.h"
 #include "common/config.h"
-#include "../network_manager/network_manager.h"
+#include "../embarlet/req_queue.h"
+#include "../embarlet/ack_queue.h"
 
 namespace Embarcadero{
-
-class NetworkManager;
 
 class DiskManager{
 	public:
 		DiskManager(std::shared_ptr<AckQueue> ack_queue, std::shared_ptr<ReqQueue> req_queue, int num_io_threads=NUM_DISK_IO_THREADS);
 		~DiskManager();
-		void SetNetworkManager(NetworkManager* network_manager){
-			network_manager_ = network_manager;
-		}
 
 	private:
 		void Disk_io_thread();
@@ -26,8 +22,6 @@ class DiskManager{
 		std::vector<std::thread> threads_;
 		std::shared_ptr<ReqQueue> reqQueue_;
 		std::shared_ptr<AckQueue> ackQueue_;
-
-		NetworkManager *network_manager_;
 
 		int log_fd_;
 		std::atomic<int> offset_{0};
